Split Ninja generation tests
Move pure Ninja rendering coverage into a dedicated unit test
module and leave integration/error-path coverage in a smaller
integration module.

Extract the shared Ninja integration case types and setup helper into
`test_support::ninja_gen` so the split modules stay below the file-size
limit while preserving the existing scenarios and snapshot coverage.
