--- tests/ninja_gen_tests.rs
@@ -1,546 +1,0 @@
-//! Unit tests for [`netsuke::ninja_gen`] Ninja file generation.
-//!
-//! Exercises [`netsuke::ninja_gen::generate`] and
-//! [`netsuke::ninja_gen::generate_into`] by constructing
-//! [`netsuke::ir::BuildGraph`] values directly and asserting the emitted
-//! Ninja syntax.  Snapshot tests use `insta` to guard output stability.
-//! Integration cases drive a real `ninja` binary against the generated
-//! file to verify build semantics end-to-end.
-
-use anyhow::{Context, Result, bail, ensure};
-use camino::Utf8PathBuf;
-use cap_std::{ambient_authority, fs_utf8::Dir};
-use insta::{Settings, assert_snapshot};
-use netsuke::ast::Recipe;
-use netsuke::ir::{Action, BuildEdge, BuildGraph};
-use netsuke::ninja_gen::{NinjaGenError, generate, generate_into};
-use rstest::{fixture, rstest};
-use std::process::Command;
-use tempfile::TempDir;
-use test_support::ninja;
-
-/// Define how the integration test should assert Ninja's behaviour.
-#[derive(Debug)]
-enum AssertionType {
-    FileContent(String),
-    FileExists,
-    StatusSuccess,
-}
-
-struct NinjaIntegrationCase {
-    action: Action,
-    edge: BuildEdge,
-    target_name: Utf8PathBuf,
-    ninja_args: Vec<&'static str>,
-    assertion: AssertionType,
-}
-
-/// Provide a temporary directory when Ninja is available, skipping otherwise.
-#[fixture]
-fn ninja_integration_setup() -> Option<TempDir> {
-    match ninja::ninja_integration_workspace() {
-        Ok(dir) => Some(dir),
-        Err(err) => {
-            tracing::warn!("skipping test: {err}");
-            None
-        }
-    }
-}
-
-#[rstest]
-#[case::phony_target_runs_command(
-    Action {
-        recipe: Recipe::Command { command: "true".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    BuildEdge {
-        action_id: "a".into(),
-        inputs: vec![Utf8PathBuf::from("in")],
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: true,
-        always: false,
-    },
-    Utf8PathBuf::from("out"),
-    concat!(
-        "rule a\n",
-        "  command = true\n\n",
-        "build out: a in\n\n",
-    ),
-)]
-#[case::standard_build(
-    Action {
-        recipe: Recipe::Command { command: "cc -c 'a.c' 'b.c' -o 'ab.o'".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    BuildEdge {
-        action_id: "compile".into(),
-        inputs: vec![Utf8PathBuf::from("a.c"), Utf8PathBuf::from("b.c")],
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("ab.o")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: false,
-        always: false,
-    },
-    Utf8PathBuf::from("ab.o"),
-    concat!(
-        "rule compile\n",
-        "  command = cc -c 'a.c' 'b.c' -o 'ab.o'\n\n",
-        "build ab.o: compile a.c b.c\n\n",
-    ),
-)]
-#[case::complex_dependencies(
-    Action {
-        recipe: Recipe::Command { command: "true".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    BuildEdge {
-        action_id: "b".into(),
-        inputs: vec![Utf8PathBuf::from("in")],
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("out"), Utf8PathBuf::from("log")],
-        implicit_outputs: vec![Utf8PathBuf::from("out.d")],
-        order_only_deps: vec![Utf8PathBuf::from("stamp")],
-        phony: false,
-        always: false,
-    },
-    Utf8PathBuf::from("out"),
-    concat!(
-        "rule b\n",
-        "  command = true\n\n",
-        "build out log | out.d: b in || stamp\n\n",
-    ),
-)]
-#[case::explicit_inputs_plus_implicit_deps(
-    Action {
-        recipe: Recipe::Command { command: "true".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    BuildEdge {
-        action_id: "b".into(),
-        inputs: vec![Utf8PathBuf::from("in")],
-        implicit_deps: vec![Utf8PathBuf::from("dep")],
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: false,
-        always: false,
-    },
-    Utf8PathBuf::from("out"),
-    concat!(
-        "rule b\n",
-        "  command = true\n\n",
-        "build out: b in | dep\n\n",
-    ),
-)]
-#[case::implicit_deps_without_explicit_inputs(
-    Action {
-        recipe: Recipe::Command { command: "true".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    BuildEdge {
-        action_id: "b".into(),
-        inputs: Vec::new(),
-        implicit_deps: vec![Utf8PathBuf::from("dep")],
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: false,
-        always: false,
-    },
-    Utf8PathBuf::from("out"),
-    concat!(
-        "rule b\n",
-        "  command = true\n\n",
-        "build out: b | dep\n\n",
-    ),
-)]
-#[case::all_dependency_classes(
-    Action {
-        recipe: Recipe::Command { command: "true".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    BuildEdge {
-        action_id: "b".into(),
-        inputs: vec![Utf8PathBuf::from("in")],
-        implicit_deps: vec![Utf8PathBuf::from("dep")],
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: vec![Utf8PathBuf::from("stamp")],
-        phony: false,
-        always: false,
-    },
-    Utf8PathBuf::from("out"),
-    concat!(
-        "rule b\n",
-        "  command = true\n\n",
-        "build out: b in | dep || stamp\n\n",
-    ),
-)]
-#[case::phony_action_with_implicit_deps(
-    Action {
-        recipe: Recipe::Command { command: "true".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    BuildEdge {
-        action_id: "phony".into(),
-        inputs: Vec::new(),
-        implicit_deps: vec![Utf8PathBuf::from("dep")],
-        explicit_outputs: vec![Utf8PathBuf::from("phony_action")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: true,
-        always: false,
-    },
-    Utf8PathBuf::from("phony_action"),
-    concat!(
-        "rule phony\n",
-        "  command = true\n\n",
-        "build phony_action: phony | dep\n\n",
-    ),
-)]
-fn generate_ninja_scenarios(
-    #[case] action: Action,
-    #[case] edge: BuildEdge,
-    #[case] target_path: Utf8PathBuf,
-    #[case] expected: &str,
-) -> Result<()> {
-    let mut graph = BuildGraph::default();
-    graph.actions.insert(edge.action_id.clone(), action);
-    graph.targets.insert(target_path, edge);
-
-    let ninja = generate(&graph)?;
-    ensure!(
-        ninja == expected,
-        "generated ninja manifest did not match expectation"
-    );
-    Ok(())
-}
-
-#[rstest]
-fn generate_empty_graph() -> Result<()> {
-    let graph = BuildGraph::default();
-    let ninja = generate(&graph)?;
-    ensure!(ninja.is_empty(), "expected empty ninja manifest");
-    Ok(())
-}
-
-#[rstest]
-fn generate_multiline_script_snapshot() -> Result<()> {
-    let mut graph = BuildGraph::default();
-    graph.actions.insert(
-        "script".into(),
-        Action {
-            recipe: Recipe::Script {
-                script: "echo one\necho two".into(),
-            },
-            description: None,
-            depfile: None,
-            deps_format: None,
-            pool: None,
-            restat: false,
-        },
-    );
-    graph.targets.insert(
-        Utf8PathBuf::from("out"),
-        BuildEdge {
-            action_id: "script".into(),
-            inputs: Vec::new(),
-            implicit_deps: Vec::new(),
-            explicit_outputs: vec![Utf8PathBuf::from("out")],
-            implicit_outputs: Vec::new(),
-            order_only_deps: Vec::new(),
-            phony: false,
-            always: false,
-        },
-    );
-    graph.default_targets.push(Utf8PathBuf::from("out"));
-
-    let ninja = generate(&graph)?;
-    let mut settings = Settings::new();
-    settings.set_snapshot_path(concat!(
-        env!("CARGO_MANIFEST_DIR"),
-        "/tests/snapshots/ninja",
-    ));
-    settings.bind(|| {
-        assert_snapshot!("multiline_script_ninja", &ninja);
-    });
-    ensure!(
-        ninja.contains("printf %b") && ninja.contains("\\n"),
-        "script should use printf %b with encoded newlines"
-    );
-    Ok(())
-}
-
-/// Integration scenarios to confirm Ninja executes commands correctly.
-#[rstest]
-#[case::multiline_script_valid(NinjaIntegrationCase {
-    action: Action {
-        recipe: Recipe::Script { script: "echo one\necho two".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    edge: BuildEdge {
-        action_id: "script".into(),
-        inputs: Vec::new(),
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: false,
-        always: false,
-    },
-    target_name: Utf8PathBuf::from("out"),
-    ninja_args: vec!["-n"],
-    assertion: AssertionType::StatusSuccess,
-})]
-#[case::script_with_percent(NinjaIntegrationCase {
-    action: Action {
-        recipe: Recipe::Script { script: "echo 100% > out".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    edge: BuildEdge {
-        action_id: "percent".into(),
-        inputs: Vec::new(),
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: false,
-        always: false,
-    },
-    target_name: Utf8PathBuf::from("out"),
-    ninja_args: vec!["out"],
-    assertion: AssertionType::FileContent("100%".into()),
-})]
-#[case::script_with_backtick(NinjaIntegrationCase {
-    action: Action {
-        recipe: Recipe::Script { script: "echo `echo hi` > out".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    edge: BuildEdge {
-        action_id: "tick".into(),
-        inputs: Vec::new(),
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: false,
-        always: false,
-    },
-    target_name: Utf8PathBuf::from("out"),
-    ninja_args: vec!["out"],
-    assertion: AssertionType::FileContent("hi".into()),
-})]
-#[case::phony_action_executes_command(NinjaIntegrationCase {
-    action: Action {
-        recipe: Recipe::Command { command: "touch action-called.txt".into() },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    },
-    edge: BuildEdge {
-        action_id: "hello".into(),
-        inputs: Vec::new(),
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("say-hello")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: true,
-        always: false,
-    },
-    target_name: Utf8PathBuf::from("action-called.txt"),
-    ninja_args: vec!["say-hello"],
-    assertion: AssertionType::FileExists,
-})]
-fn ninja_integration_tests(
-    ninja_integration_setup: Option<TempDir>,
-    #[case] case: NinjaIntegrationCase,
-) -> Result<()> {
-    let Some(dir) = ninja_integration_setup else {
-        return Ok(());
-    };
-    let dir_path = Utf8PathBuf::from_path_buf(dir.path().to_path_buf())
-        .map_err(|path| anyhow::anyhow!("temp dir path {:?} is not UTF-8", path))?;
-
-    let NinjaIntegrationCase {
-        action,
-        edge,
-        target_name,
-        ninja_args,
-        assertion,
-    } = case;
-
-    let output = edge
-        .explicit_outputs
-        .first()
-        .context("edge should have at least one explicit output")?
-        .clone();
-    if !matches!(&assertion, AssertionType::FileExists) {
-        ensure!(
-            output == target_name,
-            "expected edge output '{}' to match test target '{target_name}'",
-            output
-        );
-    }
-    let mut graph = BuildGraph::default();
-    graph.actions.insert(edge.action_id.clone(), action);
-    graph.targets.insert(output.clone(), edge);
-    graph.default_targets.push(output);
-
-    let ninja = generate(&graph)?;
-    let handle = Dir::open_ambient_dir(&dir_path, ambient_authority())
-        .with_context(|| format!("open ambient dir for temp workspace at {dir_path}"))?;
-    handle
-        .write("build.ninja", ninja.as_bytes())
-        .context("write ninja build file")?;
-    let status = Command::new("ninja")
-        .args(&ninja_args)
-        .current_dir(dir_path.as_std_path())
-        .status()
-        .context("invoke ninja")?;
-
-    match assertion {
-        AssertionType::StatusSuccess => {
-            ensure!(status.success(), "ninja invocation should succeed");
-        }
-        AssertionType::FileExists => {
-            ensure!(status.success(), "ninja invocation should succeed");
-            let exists = handle
-                .try_exists(target_name.as_str())
-                .with_context(|| format!("check existence of {target_name}"))?;
-            ensure!(
-                exists,
-                "expected {} to exist after ninja invocation",
-                &target_name
-            );
-        }
-        AssertionType::FileContent(expected) => {
-            ensure!(status.success(), "ninja invocation should succeed");
-            let content = handle
-                .read_to_string(target_name.as_str())
-                .with_context(|| format!("read target file {target_name}"))?;
-            ensure!(
-                content.trim() == expected,
-                "expected file content '{}', got '{}'",
-                expected,
-                content.trim()
-            );
-        }
-    }
-    Ok(())
-}
-
-#[rstest]
-fn errors_when_action_missing() -> Result<()> {
-    let mut graph = BuildGraph::default();
-    let edge = BuildEdge {
-        action_id: "missing".into(),
-        inputs: Vec::new(),
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: false,
-        always: false,
-    };
-    graph.targets.insert(Utf8PathBuf::from("out"), edge);
-    let Err(err) = generate(&graph) else {
-        bail!("expected missing action to error");
-    };
-    ensure!(
-        matches!(err, NinjaGenError::MissingAction { ref id, .. } if id == "missing"),
-        "unexpected error variant: {err:?}"
-    );
-    Ok(())
-}
-
-#[rstest]
-fn generate_format_error() -> Result<()> {
-    use std::fmt::{self, Write};
-
-    struct FailWriter;
-    impl Write for FailWriter {
-        fn write_str(&mut self, _: &str) -> fmt::Result {
-            Err(fmt::Error)
-        }
-    }
-
-    let action = Action {
-        recipe: Recipe::Command {
-            command: "true".into(),
-        },
-        description: None,
-        depfile: None,
-        deps_format: None,
-        pool: None,
-        restat: false,
-    };
-    let edge = BuildEdge {
-        action_id: "a".into(),
-        inputs: Vec::new(),
-        implicit_deps: Vec::new(),
-        explicit_outputs: vec![Utf8PathBuf::from("out")],
-        implicit_outputs: Vec::new(),
-        order_only_deps: Vec::new(),
-        phony: false,
-        always: false,
-    };
-    let mut graph = BuildGraph::default();
-    graph.actions.insert("a".into(), action);
-    graph.targets.insert(Utf8PathBuf::from("out"), edge);
-
-    let mut writer = FailWriter;
-    let Err(err) = generate_into(&graph, &mut writer) else {
-        bail!("expected format error when writer fails");
-    };
-    ensure!(
-        matches!(err, NinjaGenError::Format { .. }),
-        "unexpected error: {err:?}"
-    );
-    Ok(())
-}
